[gdb/testsuite] Compile ada with -lgnarl_pic and -lgnat_pic if required
On openSUSE Leap 15.2, when running the gdb.ada testsuite with target board
unix/-fPIE/-pie, I get a number of FAILs.

The problem is that although due to commit abcf2cc85a3 "[gdb/testsuite] Fix
ada tests with -fPIE/-pie" we try to compile say, hello.adb like so:
...
$ gnatmake -fPIE -largs -pie -margs hello.adb
...
this is not sufficient, because gnatlink is trying to link in libgnat.a, which
is not Position Independent Code.

This issue has been filed as gcc PR ada/87936 - "gnatlink fails with -pie".

Work around this issue by compiling with _pic versions of lgnarl and lgnat:
...
$ gnatmake -fPIE -largs -pie -lgnarl_pic -lgnat_pic -margs hello.adb
...
if that is required to make hello.adb compile.

Tested on x86_64-linux.

gdb/testsuite/ChangeLog:

2019-10-10  Tom de Vries  <tdevries@suse.de>

	PR testsuite/24888
	* lib/ada.exp (gdb_simple_compile_ada): New proc.
	(calculating_ada_needs_libs_pic_suffix): New global, initialized to 0.
	(gdb_ada_needs_libs_pic_suffix): New caching proc.
	(target_compile_ada_from_dir): Append -largs -lgnarl_pic -lgnat_pic
	-margs to multilib_flags if required.
	(gdb_compile_ada_1): Factor out of ...
	(gdb_compile_ada): ... here.

Change-Id: I3e1e40bd46236b45e2d7808c1cd744a075d4a148