Introduce DWARF abbrev cache
The replacement for the DWARF psymbol reader works in a somewhat
different way.  The current reader reads and stores all the DIEs that
might be interesting.  Then, if it is missing a DIE, it re-scans the
CU and reads them all.  This approach is used for both intra- and
inter-CU references.

I instrumented the partial DIE hash to see how frequently it was used:

    [  0] -> 1538165
    [  1] ->    4912
    [  2] ->   96102
    [  3] ->     175
    [  4] ->     244

That is, most DIEs are never used, and some are looked up twice -- but
this is just an artifact of the implementation of
partial_die_info::fixup, which may do two lookups.

Based on this, the new implementation doesn't try to store any DIEs,
but instead just re-scans them on demand.  In order to do this,
though, it is convenient to have a cache of DWARF abbrevs.  This way,
if a second CU is needed to resolve an inter-CU reference, the abbrevs
for that CU need only be computed a single time.

2021-05-24  Tom Tromey  <tom@tromey.com>

	* dwarf2/abbrev.h (struct abbrev_table) <section>: New member.
	<abbrev_table>: Add 'sect' parameter.
	* dwarf2/abbrev.c (abbrev_table::abbrev_table): Add 'sect'
	parameter.
	(abbrev_table::read): Update.
	* dwarf2/abbrev-cache.h: New file.
	* dwarf2/abbrev-cache.c: New file.
	* Makefile.in (COMMON_SFILES): Add abbrev-cache.c.