fixup-gdb-6.5-sharedlibrary-path
[gdb/testsuite] Fixup gdb.threads/tls-sepdebug.exp

On factory, due to --as-needed we end up with a tls-sepdebug-main without
dependency on tls-sepdebug-shared.so:
...
$ ldd tls-sepdebug-main
        linux-vdso.so.1 (0x00007fffe7fcf000)
        libc.so.6 => /lib64/libc.so.6 (0x00007f205f05f000)
        /lib64/ld-linux-x86-64.so.2 (0x00007f205f232000)
...

Fix this by actually using the variable var (defined in
tls-sepdebug-shared.so) in tls-sepdebug-main.

Likewise, we end up without a dependency on libpthread.so.0, which we need to
read tls vars.  Add a call to pthread_testcancel, as in
gdb.threads/tls-var-main.c