Statically examine abbrev properties
The new DIE scanner works more or less along the lines indicated by
the text for the .debug_names section, disregarding the bugs in the
specification.

While working on this, I noticed that whether a DIE is interesting is
a static property of the DIE's abbrev.  It also turns out that many
abbrevs imply a static size for the DIE data, and additionally that
for many abbrevs, the sibling offset is stored at a constant offset
from the start of the DIE.

This patch changes the abbrev reader to analyze each abbrev and stash
the results on the abbrev.  These combine to speed up the new indexer.
If the "interesting" flag is false, GDB knows to skip the DIE
immediately.  If the sibling offset is statically known, skipping can
be done without reading any attributes; and in some other cases, the
DIE can be skipped using simple arithmetic.

2021-05-24  Tom Tromey  <tom@tromey.com>

	* dwarf2/abbrev.h (struct abbrev_info) <tag>: Use ENUM_BITFIELD.
	<has_children>: Now bool.
	<interesting, size_if_constant, sibling_offset>: New members.
	* dwarf2/abbrev.c (tag_interesting_for_index): New function.
	(abbrev_table::read): Examine attributes and compute
	interesting-ness.