"Finalize" the DWARF index in the background
After scanning the CUs, the DWARF indexer merges all the data into a
single hash table, canonicalizing C++ names as it proceeds.  While not
necessarily single-threaded, this process is currently done in just
one thread, to keep memory costs lower.

However, this work is all done without reference to any data outside
of the indexes.  This patch improves the apparent performance of GDB
by moving it to the background.  All uses of the index are then made
to wait for this process to complete.

In our ongoing example, this reduces the scanning time on (an
unchanging copy of) gdb itself to 0.173937 (wall).  Recall that before
this patch, the time was 0.668923; and psymbol reader does this in
1.598869.  That is, at the end of this series, we see about a 10x
speedup.

2021-05-24  Tom Tromey  <tom@tromey.com>

	* dwarf2/cooked-index.h (~cooked_index_vector): New.
	(cooked_index_vector::find): Wait for m_future.
	(cooked_index_vector::m_future): New member.
	* dwarf2/cooked-index.c
	(cooked_index_vector::cooked_index_vector): Initialize m_future.
	Call finalize in background.
	(cooked_index_vector::traverse): Wait for m_future.