Refactor build_type_psymtabs_reader
The new DWARF scanner needs to save the entire cutu_reader object, not
just parts of it.  In order to make this possible, this patch
refactors build_type_psymtabs_reader.  This change is done separately
because it is easy to review in isolation and it helps make the later
patches smaller.

gdb/ChangeLog
2021-06-18  Tom Tromey  <tom@tromey.com>

	* dwarf2/read.c (process_psymtab_comp_unit): Update.
	(build_type_psymtabs_reader): Remove info_ptr and type_unit_die
	parameters.  Change reader to cutu_reader.
	(build_type_psymtabs, process_skeletonless_type_unit): Update.