include, gdb: fix -Wswitch build errors with gcc 4.8
I started seeing these strange errors when building with gcc 4.8:

      CXX    ada-tasks.o
    /home/smarchi/src/binutils-gdb/gdb/ada-tasks.c: In function ‘void read_known_tasks()’:
    /home/smarchi/src/binutils-gdb/gdb/ada-tasks.c:998:10: error: enumeration value ‘ADA_TASKS_UNKNOWN’ not handled in switch [-Werror=switch]
       switch (data->known_tasks_kind)
              ^

It is caused by commit 06de25b7af21 ("gdb: introduce
target_waitkind_str, use it in target_waitstatus::to_string"), which
introduced some pragmas to enable -Wswitch locally.  That shouldn't
affect the rest of the code, since we are using push and pop around
that.  It looks like gcc 4.8 (and 4.9)'s diagnostic push/pop is not
reliable, as it doesn't cause a problem with later versions.

Work around this by not enabling -Wswitch for GCC < 5.  This makes the
code a bit ugly, it would be nice to find a good way to factor this out,
especially if we want to re-use it elsewhere.  But for now, I just want
to un-break the build.

Note that this code (already as it exists in master today) enables
-Wswitch at the error level even if --disable-werror is passed.  It
shouldn't be a problem, as it's not like a new enumerator will appear
out of nowhere.  So it shouldn't cause a spurious build error in the
future.  Still, for correctness, we would ideally want to ask the
compiler to enable -Wswitch at its default level (as if the user had
passed -Wswitch on the command-line).  There doesn't seem to be a way to
do this.

Change-Id: I50d66b348bf83de099c3c0879e2eb352d60540ba